Add registers and implement basic write
Note: spi_cmds will not compile until noted in commit msgs. Added
registers in such way that it is easy to use them from both
userspace/kernelspace if needed, as they have encoded bank in them.
Basic write was written, but until all write/read procedure for spi are
in place these are not really usable.